agents/recipe_agent/pyproject.toml
[project]
name = "recipe-agent"
version = "0.1.0"
description = ""
requires-python = ">=3.12"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "~3.12"
flask = "^3.0.3"
google-generativeai = "^0.8.4"
python-dotenv = "^1.0.1"
waitress = "^3.0.2"
